requirements.txt
Django==1.6.8
Pillow==2.7.0
PyPDF2==1.24
South==1.0.2
argparse==1.2.1
django-classy-tags==0.6.1
django-cms==3.0.9
django-easy-pdf==0.1.0
django-mptt==0.6.1
django-sekizai==0.8.1
djangocms-admin-style==0.2.5
djangocms-column==1.5
djangocms-style==1.5
djangocms-text-ckeditor==2.4.3
html5lib==0.999
reportlab==3.1.44
six==1.9.0
wsgiref==0.1.2
xhtml2pdf==0.0.6

